requirements.txt
streamlit==1.38.0
streamlit-autorefresh
pyarrow==15.0.2
protobuf<6
grpcio<1.64
httpx==0.27.2
pandas
openai
yfinance
plotly
requests
matplotlib
email-validator
schedule
python-dotenv
